Sprint #5

+++ src/rutas/usuariosGestion/eliminarApi.js
@@ -6,28 +6,43 @@ eliminarApi.post("/eliminar", async function(req, res){
     console.log("entró a eliminar")
     try{
         const { cedula } = req.body
+        const { validar } = req.body
         const c = await usuariosModel.findOne({ cedula })   
-        console.log(c)
+        console.log(cedula)
+        console.log(validar)
         if(c){
+
             if(c.rol != "cliente"){
-                console.log("entró a verificar el estado");
-                console.log(cedula)
-                console.log(c.rol)
-                const d = await usuariosModel.deleteOne({ cedula })
-                console.log(d)
-                return res.status(200).send({
-                    estado: "ok",
-                    msg: "Usuario Eliminado"
-                })//se puede eliminar
+
+                if(validar != cedula){
+
+                    console.log("entró a verificar el estado");
+                    console.log(cedula)
+                    console.log(c.rol)
+                    const d = await usuariosModel.deleteOne({ cedula })
+                    console.log(d)
+                    return res.status(200).send({
+                        estado: "ok",
+                        msg: "Usuario Eliminado"
+                    })//se puede eliminar
+
+                }else{
+                    return res.status(200).send({
+                        estado: "igual"
+                    })
+
+                }
+               
             }else{
                 return res.status(500).send({
                     estado: "error",
-                    msg: "este usuario no se puede eliminar"
+                    msg: "No se pueden eliminar clientes"
                 })    
                 //no se puede eliminar
             }
 
         } else if(!c){
+
             return res.status(500).send({
                 estado: "error",
                 msg: "cedula no encontrada en plataforma"
